+++ lib/transformer.dart
@@ -1,0 +1,9 @@
+library transformer;
+
+import 'dart:async';
+
+import 'package:barback/barback.dart';
+import 'package:analyzer/analyzer.dart';
+import 'package:analyzer/src/generated/scanner.dart';
+
+part 'src/transformer/dartemis_transformer.dart';