+++ test/dartemis/transformer/_all_tests.dart
@@ -1,0 +1,7 @@
+library transformer._all_tests;
+
+import 'dartemis_transformer_test.dart' as dartemis_transformer;
+
+void main() {
+  dartemis_transformer.main();
+}