+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### API Changes
 * deprecated `ComponentMapper` use `Mapper` instead
 * deprecated `ComponentMapper#get(Entity)`, use `Mapper[Entity]` instead
+* deprecated `dartemis_mirror#World`, use the transformer instead
+### Enhancements
+* created a code generating transformer to insert `Manager`, `System`s and `Mapper`s into `Manager` and `System` instances 
 ### Bugfixes
 * DelayedEntityProcessingSystem keeps running until all current entities have expired
 ### Internal